+++ catroid/src/androidTest/java/org/catrobat/catroid/test/physics/PhysicsWorldTest.java
@@ -44,7 +44,7 @@ public class PhysicsWorldTest extends AndroidTestCase {
 		GdxNativesLoader.load();
 	}
 
-	private static final String TAG = PhysicsObjectTest.class.getSimpleName();
+	private static final String TAG = PhysicsWorldTest.class.getSimpleName();
 	private PhysicsWorld physicsWorld;
 	private World world;
 	private Map<Sprite, PhysicsObject> physicsObjects;
